movieMela/requirement.txt
dj-database-url==0.5.0
Django==2.2.5
django-filer==1.5.0
django-filter==2.2.0
django-js-asset==1.2.2
django-mptt==0.10.0
django-polymorphic==2.0.3
djangorestframework==3.10.3
djangorestframework-composed-permissions==0.2.1
djangorestframework-simplejwt==4.3.0
easy-thumbnails==2.6
gunicorn==19.9.0
Markdown==3.1.1
Pillow==6.1.0
psycopg2-binary==2.8.3
PyJWT==1.7.1
pytz==2019.2
shell==1.0.1
sqlparse==0.3.0
Unidecode==1.0.23
whitenoise==4.1.4
